bugfix: ensure unit tests are being properly executing
Naming changed somehow due to renaming of the base class probably. This changes back the naming so that the tests are actually executed.
